xml/parsing/manakai/eldecls-perefs-1.dat

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT %fuga;EMPTY>
#errors
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT xy >
| <{} a>

#data
<!DOCTYPE a SYSTEM "http://a/"><a />
#resource http://a/
<!ENTITY % fuga "xy">
<!ELEMENT %fuga; EMPTY>
#errors
#document
| <!DOCTYPE a "" "http://a/">
|   <!ELEMENT xy >
| <{} a>
